Review Avax tests, fixing token amounts to test ERC4626 tokens in the extremes (#57)
* Note about aAAVE token

* Note about aAUSD token

* Change amount of BTC to test

* Fix DAI amount to test

* Fix FRAX amount

* Fix token amounts

* Fix token amount

* Fix amount and add note about aUSDT token

* Updates test to avoid minting issues

Updates the test to use a different donor address and adjusts the amount donated to avoid minting failures when depositing into Aave on Avalanche.

* Updates WETH donor and donation amount

Updates the WETH donor address and the amount to donate
in the Avalanche Aave WETH test to align with current
test environment and requirements.

* Updates test parameters for Avalanche Euler

Updates the underlying donor addresses and amount to donate
for the ERC4626 Avalanche Euler tests.

This ensures the tests work correctly with the current state
of the Avalanche network and Euler protocol. Specifically,
updates addresses to accounts with sufficient funds for testing
and adjust the donation amounts to prevent failures due to liquidity limits.